long int for filesize

+++ mreceive.h
@@ -1,5 +1,5 @@
 struct mheader 
 {
 	char filename[128];
-	int length;
+	long length;
 };
